+++ lab1/Linechart.pde
@@ -6,13 +6,13 @@ public class Linechart {
     int radius = 5;
 
     String intersect (int mousex, int mousey) {
-        int ratio = (topyaxis.y - origin.y) / maxY;
+        float ratio = float((topyaxis.y - origin.y)) / maxY;
         int sectionWidth = abs(((rightxaxis.x - origin.x) / datapoints.length));
         strokeWeight(sectionWidth * 0.8);
         
         for (int i = 0; i < datapoints.length; i++) {
             int x = origin.x + sectionWidth * i + sectionWidth / 2 + int(sectionWidth * 0.1);
-            int y = datapoints[i].count * ratio + origin.y + 4;
+            int y = int(datapoints[i].count * ratio) + origin.y + 4;
             
             if ((mousex - x) * (mousex - x) + (mousey - y) * (mousey - y) < (radius + 1) * (radius + 1)) {
                 return "(" + datapoints[i].fruit + ", " + datapoints[i].count + ")";
@@ -39,13 +39,13 @@ public class Linechart {
 
     void draw () {
         strokeWeight(2);
-        int ratio = (topyaxis.y - origin.y) / maxY;
+        float ratio = float((topyaxis.y - origin.y)) / maxY;
         int sectionWidth = abs(((rightxaxis.x - origin.x) / datapoints.length));
-        Point prev = new Point(origin.x + sectionWidth / 2, datapoints[0].count * ratio + origin.y);
+        Point prev = new Point(origin.x + sectionWidth / 2, int(datapoints[0].count * ratio) + origin.y);
         drawCircle(prev.x, prev.y, 10);
         for (int i = 1; i < datapoints.length; i++) {
             int x = origin.x + sectionWidth * i + sectionWidth / 2 + int(sectionWidth * 0.1);
-            int y = datapoints[i].count * ratio + origin.y;
+            int y = int(datapoints[i].count * ratio) + origin.y;
             line(prev.x, prev.y, x, y);
             prev.setXY(x, y);
             drawCircle(prev.x, prev.y, radius * 2);
